Applying updates from Tore, cleaning up interfaces.
git-svn-id: svn+ssh://src.dsl.bbn.com/svn/openmap/trunk@1961 f68c23cc-7f87-4fbc-acfa-d9760b90acc1

+++ src/openmap/com/bbn/openmap/proj/coords/CoordinateReferenceSystem.java
@@ -230,7 +230,7 @@ public class CoordinateReferenceSystem {
 
     public void prepareProjection(GeoProj proj) {
         // TODO: do we need this??
-        proj.setPlanetRadius((float) ellipsoid.radius);
+        proj.setPlanetRadius(ellipsoid.radius);
     }
 
     /**
